Update 0.4.0 Code
I have been working on the new end screen for #MQO and with useful information (e.g. your name and and score) and can be printed looking good and can be used as a certificate reward! There is also a new score target system in place as a real challenge for the player or just use endless mode to get the biggest score!